0.5.3 - added optional CORS support for clientside JS API access
- __New Features:__
    + Optional CORS support from [Michael Bumann](https://twitter.com/bumi) (thanks!).
- __Improvements:__
    + Bumped `hiredis` to 0.4.1 for latest iojs compat.

Although DNSChain isn't intended to be used from JavaScript within the web browser (unless it's via a browser extension), this releases makes that possible. Doing so allows client-side JavaScript to access blockchain data via the RESTful API, and "secured" (not really) by X.509 and HTTPS.

To enable this, add the following to your DNSChain config:

```
[http]

cors=true
```

Keep in mind that enabling this on a public web server means that anyone on the Internet can use it to access blockchain data. You may therefore want to test and adjust the request rate limiting settings as necessary.

Thanks to Michael Bumann for this addition!